Kafka installation with more options for KAFKA_HOME and version settings

+++ lib/bash/debian/kafka_bin.sh
@@ -1,11 +1,35 @@
 #!/bin/bash
-
 # See also https://gist.github.com/monkut/07cd1618102cbae8d587811654c92902
 # See also https://www.digitalocean.com/community/tutorials/how-to-install-apache-kafka-on-ubuntu-14-04
 
 # Dependencies (are not installed by this script)
 # - java (runtime)
 # - zookeeper (runtime)
+
+# Help info
+
+HELP=$1
+if [ "$HELP" == '-h' -o "$HELP" == '--help' ]; then
+    echo "$0 [SCALA_VERSION] [KAFKA_VERSION]"
+    exit
+fi
+
+# Versions
+
+SCALA_VER=$1
+if [ -z "$SCALA_VER" ]; then
+    SCALA_VER=2.11
+fi
+
+KAFKA_VER=$2
+if [ -z "$KAFKA_VER" ]; then
+    KAFKA_VER=0.11.0.0
+fi
+
+KAFKA_BIN="kafka_${SCALA_VER}-${KAFKA_VER}"
+KAFKA_TAR="${KAFKA_BIN}.tgz"
+
+echo "Installing $KAFKA_BIN"
 
 # Create a user for Kafka
 # - this requires an interactive shell, so it's commented out here.
@@ -16,26 +40,33 @@
 #adduser kafka sudo
 #su - kafka
 
-# The KAFKA_HOME is a symlink to a Kafka version directory
+# ---
+# Setup and use the KAFKA_HOME directory
+# - it is a symlink to a versioned Kafka installation
+# - it is stored permanently in /etc/profile.d/kafka.sh
+
+if [ -z "$KAFKA_HOME" ]; then
+    export KAFKA_HOME='/opt/kafka'
+fi
 if [ ! -f /etc/profile.d/kafka.sh ]; then
-    echo 'export KAFKA_HOME=/usr/local/kafka' | sudo tee -a /etc/profile.d/kafka.sh
+    echo "export KAFKA_HOME=${KAFKA_HOME}"       | sudo tee -a /etc/profile.d/kafka.sh
+    echo "export PATH=${PATH}:${KAFKA_HOME}/bin" | sudo tee -a /etc/profile.d/kafka.sh
 fi
 source /etc/profile.d/kafka.sh
 export KAFKA_INSTALL_DIR=$(dirname ${KAFKA_HOME})
 
+# ---
 # Check if it's installed already
+KAFKA_PATH="${KAFKA_INSTALL_DIR}/${KAFKA_BIN}"
 
-SCALA_VER=2.11
-VER=0.11.0.0
-BIN="kafka_${SCALA_VER}-${VER}"
-TAR="${BIN}.tgz"
-
-if ls -d ${KAFKA_INSTALL_DIR}/${BIN} > /dev/null 2&>1; then
-    echo "Found Kafka installed already in:"
-    ls -d ${KAFKA_INSTALL_DIR}/${BIN}
+if [ -d ${KAFKA_PATH} ]; then
+    echo "Found Kafka installed already in: ${KAFKA_PATH}"
+    echo "Updating the ${KAFKA_HOME} symlink to use ${KAFKA_PATH}"
+    rm -f ${KAFKA_HOME}
+    ln -s ${KAFKA_PATH} ${KAFKA_HOME}
     exit 0
 else
-    echo "Did not find Kafka installed in: ${KAFKA_INSTALL_DIR}/${BIN}"
+    echo "Did not find Kafka installed in: ${KAFKA_INSTALL_DIR}/${KAFKA_BIN}"
 fi
 
 
@@ -43,17 +74,21 @@ fi
 # Download binary release and install it
 mkdir -p ~/Downloads/apache-kafka
 cd ~/Downloads/apache-kafka
-wget -q -nc https://archive.apache.org/dist/kafka/${VER}/${TAR}
+wget -q -nc https://archive.apache.org/dist/kafka/${KAFKA_VER}/${KAFKA_TAR}
 
-#http://mirror.olnevhost.net/pub/apache/kafka/${VER}/${TAR}
+# check the success of the download
+if [ $? -ne 0 ]; then
+    echo "Failed to download ${KAFKA_TAR}"
+    exit 1
+fi
+
 
 # ---
 # Install to KAFKA_INSTALL_DIR and create symlink to KAFKA_HOME
-tar -zxf ${TAR} -C ${KAFKA_INSTALL_DIR}/
-# remove any existing symlink (it might not exist)
+tar -zxf ${KAFKA_TAR} -C ${KAFKA_INSTALL_DIR}/
+# create or update the KAFKA_HOME symlink
 rm -f ${KAFKA_HOME}
-# create the KAFKA_HOME symlink
-ln -s ${KAFKA_INSTALL_DIR}/${BIN} ${KAFKA_HOME}
+ln -s ${KAFKA_PATH} ${KAFKA_HOME}
 
 
 # ---
